UI and Interaction Improvements + Semi tab system
Implemented ICONS where the text originally was for back, forward, home, and reload. Added a small reload animation to the button during loading. Implementation of a reload cancel button is almost redundant, so I'm considering binding it to a keypress like ctrl+L or something.

Added a basic text display to the homepage to not confuse it for a loading error. Still needs major work to not look absolutely awful as it's right now just plain text, I'll have to do some CSS formatting for it as well as add interactive bookmarks or other features with JS. Setting menu would also be a good idea.